+++ src/main/java/com/ldh/todolist/service/TodoService.java
@@ -13,11 +13,11 @@ public interface TodoService {
 	void updateTodo(TodoDto todoDto);
 
 	//할 일 삭제
-	void deleteTodo(int todoNo);
+	void deleteTodo(Long todoNo);
 	
 	//할 일 조회
-	TodoDto findById(int todoNo);
+	TodoDto findById(Long todoNo);
 	
 	//할 일 목록 조회(사용자 번호)
-	List<TodoDto> findByUsersNo(int usersNo);
+	List<TodoDto> findByUsersNo(Long usersNo);
 }
